Windows: Fix QtWebEngine SIGTRAP
In Windows we had an issue where bran would get a SIGTRAP on start, with no
output or window being shown.

We never got to the root cause of the issue, the debug only determined it
happened in QtWebEngine @GetEventHandler. In the end I think we were dealing
with multiple issues (or the same issue at multiple places) an all resulted in
a SIGTRAP. With some trial and error we ended up doing the following:

- Allocate in the HEAP rather than the stack (QApplication in particular)
- Delay the invocation of QWebEngineProfile::setHttpUserAgent(), after the
  creation of the view/page.